v2.4.0 rc1 AmorLib dependency, Arrayables, SetOutsideDimensionDataEvent, move log mode, started checkpoint baseline, LiveEdit for Config

+++ AWO/Utils/RandomExtensions.cs
@@ -20,6 +20,6 @@ public static class RandomExtensions
 
     public static float NextFloat(this Random rand)
     {
-        return (float)rand.NextDouble();
+        return rand.NextSingle();
     }
 }
